requirements.txt
celery==5.4.0
channels==4.1.0
drf-yasg==1.21.7
Django==4.2.14
djangorestframework==3.15.2
django-celery-beat==2.6.0
flower==2.0.1
gunicorn==22.0.0
psycopg==3.2.1
python-dotenv==1.0.1
redis==5.0.7
